Add support for Yarn PnP repos
For repos that use Yarn PnP, there is no `node_modules` directory. The
code in this extension assumes that that directory exists, and
hard-codes the location of the TypeScript SDK to path inside
`node_modules`.

This patch allows specifying LSP initialization options to override this
behavior. For Yarn PnP users, they can point to the location where the
SDK is installed after running `yarn dlx @yarnpkg/sdks`. This is
typically `.yarn/sdks/typescript/lib`.