Allow PnP users to specify TypeScript SDK location (#11)
As described in #8, there is no `node_modules` directory when working in
a Yarn PnP project. One of the assumptions this extension makes is that
the TypeScript SDK is located in `node_modules`.

This patch allows users of this extension to specify the `tsdk` location
via the LSP `initialization_options`, if required. By default, the
existing behavior is retained.

Also closes #9.

(Note, I don't think it fully fixes the issues in #8, but this is maybe
step 1?)

---------

Co-authored-by: Marshall Bowers <git@maxdeviant.com>