Improve tag auto-indent (#31)
Improves auto-indent and newline.only queries. This makes it so when
hitting enter between an open and closing tag like so
```html
<div>|</div>
```
The result is:
```html
<div>
    |
</div>
```
Instead of:
```html
<div>
|</div>
```

Even when the closing tag does not have the same name as the opening tag
(i.e. `<div></p>` still auto-indents correctly)